Move view helper functions to domain module
