+++ codespeed/tests/test_models.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import os
 
 from django.conf import settings
